Move test programs to their own subdirectory
