+++ tests/kat_kem.c
