+++ tests/kat_sig.c
