+++ tests/test_kem.c
